fix: cannot open connect wallet modal cause a has connected already
