+++ src/modals/ChainModal.tsx
@@ -2,12 +2,12 @@ import type { SingletonModalRefCreator } from '@masknet/shared-base';
 import { useChainModal } from '@rainbow-me/rainbowkit';
 import { forwardRef } from 'react';
 
-import { useRainbowModalRef } from '@/hooks/useRainbowModalRef.js';
+import { useRainbowModal } from '@/hooks/useRainbowModal.js';
 
 export const ChainModal = forwardRef<SingletonModalRefCreator>(function ChainModal(_, ref) {
     const { openChainModal, chainModalOpen } = useChainModal();
 
-    useRainbowModalRef(openChainModal, chainModalOpen, ref);
+    useRainbowModal(openChainModal, chainModalOpen, ref);
 
     return null;
 });
