+++ src/modals/ConnectWalletModal.tsx
@@ -2,12 +2,12 @@ import type { SingletonModalRefCreator } from '@masknet/shared-base';
 import { useConnectModal } from '@rainbow-me/rainbowkit';
 import { forwardRef } from 'react';
 
-import { useRainbowModalRef } from '@/hooks/useRainbowModalRef.js';
+import { useRainbowModal } from '@/hooks/useRainbowModal.js';
 
 export const ConnectWalletModal = forwardRef<SingletonModalRefCreator>(function ConnectWalletModal(_, ref) {
     const { openConnectModal, connectModalOpen } = useConnectModal();
 
-    useRainbowModalRef(openConnectModal, connectModalOpen, ref);
+    useRainbowModal(openConnectModal, connectModalOpen, ref);
 
     return null;
 });
