+++ src/modals/LoginModal.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { Dialog, Transition } from '@headlessui/react';
-import { t,Trans } from '@lingui/macro';
+import { t, Trans } from '@lingui/macro';
 import { delay, getEnumAsArray } from '@masknet/kit';
 import type { SingletonModalRefCreator } from '@masknet/shared-base';
 import { useSingletonModal } from '@masknet/shared-base-ui';
